clusterpolicy_controller: watch Nodes' GPU labels to support scale-up
This patch adds a custom Watch to the controller. The GPU operator
watches for Node events. When it happens, it checks if the GPU common
label is correctly applied and synchronized with the NFD labels.

When a mismatch is detected, the custom mapping function looks up all
the existing ClusterPolicy objects and requests a reconciliation.

Without this patch, when new nodes joined the cluster, they would
not load the Nvidia driver & toolkit until a GPU operator
reconciliation would be triggered (for instance by deleting an object
controlled by the operator).

Fix: NVIDIA/gpu-operator!81

Signed-off-by: Kevin Pouget <kpouget@redhat.com>